chore(table): rename data table
