+++ src/demo-app/demo-app-module.ts
@@ -37,13 +37,13 @@ import {PlatformDemo} from './platform/platform-demo';
 import {AutocompleteDemo} from './autocomplete/autocomplete-demo';
 import {InputDemo} from './input/input-demo';
 import {StyleDemo} from './style/style-demo';
-import {DataTableDemo} from './data-table/data-table-demo';
-import {PeopleDatabase} from './data-table/people-database';
+import {TableDemo} from './table/table-demo';
+import {PeopleDatabase} from './table/people-database';
 import {DatepickerDemo} from './datepicker/datepicker-demo';
 import {TypographyDemo} from './typography/typography-demo';
 import {ExpansionDemo} from './expansion/expansion-demo';
 import {
-  CdkDataTableModule,
+  CdkTableModule,
   FullscreenOverlayContainer,
   MdAutocompleteModule,
   MdButtonModule,
@@ -79,7 +79,7 @@ import {
   OverlayContainer,
   StyleModule
 } from '@angular/material';
-import {TableHeaderDemo} from './data-table/table-header-demo';
+import {TableHeaderDemo} from './table/table-header-demo';
 
 /**
  * NgModule that includes all Material modules that are required to serve the demo-app.
@@ -117,7 +117,7 @@ import {TableHeaderDemo} from './data-table/table-header-demo';
     MdToolbarModule,
     MdTooltipModule,
     MdNativeDateModule,
-    CdkDataTableModule,
+    CdkTableModule,
     StyleModule
   ]
 })
@@ -142,7 +142,6 @@ export class DemoMaterialModule {}
     CardDemo,
     ChipsDemo,
     CheckboxDemo,
-    DataTableDemo,
     DatepickerDemo,
     DemoApp,
     DemoAppOnPush,
@@ -177,6 +176,7 @@ export class DemoMaterialModule {}
     TableHeaderDemo,
     ToolbarDemo,
     TooltipDemo,
+    TableDemo,
     TabsDemo,
     SunnyTabContent,
     RainyTabContent,
