+++ src/demo-app/demo-app/demo-app.ts
@@ -40,7 +40,6 @@ export class DemoApp {
     {name: 'Card', route: 'card'},
     {name: 'Chips', route: 'chips'},
     {name: 'Checkbox', route: 'checkbox'},
-    {name: 'Data Table', route: 'data-table'},
     {name: 'Datepicker', route: 'datepicker'},
     {name: 'Dialog', route: 'dialog'},
     {name: 'Expansion Panel', route: 'expansion'},
@@ -62,6 +61,7 @@ export class DemoApp {
     {name: 'Slider', route: 'slider'},
     {name: 'Slide Toggle', route: 'slide-toggle'},
     {name: 'Snack Bar', route: 'snack-bar'},
+    {name: 'Table', route: 'table'},
     {name: 'Tabs', route: 'tabs'},
     {name: 'Toolbar', route: 'toolbar'},
     {name: 'Tooltip', route: 'tooltip'},
