+++ src/lib/module.ts
@@ -44,7 +44,7 @@ import {PlatformModule} from './core/platform/index';
 import {MdAutocompleteModule} from './autocomplete/index';
 import {StyleModule} from './core/style/index';
 import {MdDatepickerModule} from './datepicker/index';
-import {CdkDataTableModule} from './core/data-table/index';
+import {CdkTableModule} from './core/table/index';
 import {MdExpansionModule} from './expansion/index';
 import {MdTableModule} from './table/index';
 import {MdSortModule} from './sort/index';
@@ -88,7 +88,7 @@ const MATERIAL_MODULES = [
   PlatformModule,
   MdCommonModule,
   ObserveContentModule,
-  CdkDataTableModule,
+  CdkTableModule
 ];
 
 /** @deprecated */
