+++ src/lib/public_api.ts
@@ -21,8 +21,6 @@ export * from './button-toggle/index';
 export * from './card/index';
 export * from './chips/index';
 export * from './checkbox/index';
-export * from './core/data-table/index';
-export * from './table/index';
 export * from './datepicker/index';
 export * from './dialog/index';
 export * from './expansion/index';
